Review edits -- factor out tx matching and block helpers into own module
